requirements.txt
pandas>=2.0.3
numpy>=1.24.3
scikit-learn>=1.3.0
flask>=2.3.2
flask-cors>=4.0.0
joblib>=1.2.0
xgboost>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
PyJWT>=2.8.0